feat: add mint tests (#2)

+++ src/interfaces/IArrakisMetaLP.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 interface IArrakisMetaLP {
-    function deposit(uint64 proportion) external;
+    function deposit(uint256 proportion) external;
     function withdraw(uint24 proportion, address receiver) external returns (uint256 amount0, uint256 amount1);
     function rebalance(address[] calldata targets, bytes[] calldata payloads) external;
     function setManager(address newManager) external;
